core: define default implementation for most of smbus
This change adds a default implementation for a majority of the
smbus functions to aid in supporting additional platforms in
the future.

At present, the only consumer of these defaults is the
test code which has been cleaned up quite a bit.  Right now
most of the work is delegated to the I2CRegisterMap.  It is
expected that changes may need to be made to the MockI2CDevice
itself in the future to support new use cases that may arise.

This change also includes some test fixups for the
mpl115a2.